Fix Wrapper module & wrap the WhoRU module
